requirements.txt
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
tkinter
pandas==2.0.3
face-recognition==1.3.0
flask==2.3.3
flask-cors==4.0.0